Refactoring and typos fixed
